\name{gk_list_examples}
\alias{gk_list_examples}
\title{List Available Example Datasets}
\usage{
gk_list_examples()
}
\value{
Character vector of bundled example names.
}
\description{
Returns the list of example datasets that can be loaded with
\code{gk_load_example()}.
}
\examples{
\dontrun{
gk_list_examples()
}
}
