\name{gk_methods}
\alias{gk_methods}
\title{List Available Detection Methods}
\usage{
gk_methods()
}
\value{
Character vector of method names accepted by \code{gk_detect()}.
}
\description{
Returns the list of gait-event detection methods available in the Python backend.
}
\examples{
\donttest{
gk_methods()
}
}
